requirements.txt
aiofiles==0.8.0
aiopg==1.3.4
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
bcrypt==3.2.2
cffi==1.15.1
greenlet==1.1.2
httptools==0.4.0
jsonschema==4.14.0
multidict==6.0.2
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.4.0
pyrsistent==0.18.1
sanic==22.6.2
sanic-routing==22.3.0
simple-bcrypt==0.1.0
SQLAlchemy==1.4.40
ujson==5.4.0
uvloop==0.16.0
websockets==10.3
